Don’t include the nginx::ohai_plugin recipe by default
We install our custom `pay-nginx` package, which currently needs to be
installed using the —allow-unauthenticated flag for reasons. As there is
no way to set package installation options via node attributes we need
to install pay-nginx before including the nginx main cookbook. This is 
causing issues on Chef client first runs resulting in:

```
[2016-05-31T14:12:16+00:00] ERROR: ohai[reload_nginx] (nginx::ohai_plugin line 22) had an error: Ohai::Exceptions::AttributeNotFound: No such attribute: 'nginx'
[2016-05-31T14:12:17+00:00] FATAL: Chef::Exceptions::ChildConvergeError: Chef run process exited unsuccessfully (exit code 1)
```

The second run passes fine, but this is not ideal. I’ve tried bumping
Ohai cookbook versions which hasn’t fixed. Also we are already using the
2.7.4 version of the cookbook which is supposed to fix this:
https://github.com/miketheman/nginx/issues/408

The main Nginx cookbook is undergoing a rewrite which is not currently 
ready which makes forking this cookbook a little less bad as it is not
likely to carry on in this state anyway. We don’t need the Nginx Ohai 
attributes and I’ve wasted enough time trying to get to the bottom of
this.